+++ validation/schemas.js
@@ -1,0 +1,23 @@
+const Joi = require("joi");
+
+const postSchema = Joi.object({
+  name: Joi.string().min(3).max(30).required(),
+  email: Joi.string()
+    .email({ minDomainSegments: 2, tlds: { allow: ["com", "net"] } })
+    .required(),
+  phone: Joi.string().required(),
+});
+
+const putSchema = Joi.object({
+  name: Joi.string().min(3).max(30),
+  email: Joi.string().email({
+    minDomainSegments: 2,
+    tlds: { allow: ["com", "net"] },
+  }),
+  phone: Joi.string(),
+}).min(1);
+
+module.exports = {
+  postSchema,
+  putSchema,
+};
